p2p-securemsg/requirements.txt
# P2P Secure Messaging Dependencies
# Core cryptography and security
cryptography>=41.0.0
PyNaCl>=1.5.0

# Async IO and networking
asyncio

# System utilities
psutil>=5.9.0

# Rich CLI interface
rich>=13.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0 